+++ test/unit/gemm/device/testbed_gemm_with_reduction.h
@@ -362,7 +362,7 @@ struct TestbedGemmWithReduction {
     // Determine SMEM requirements and waive if not satisfied
     //
 
-    int smem_size = int(sizeof(typename Gemm::GemmKernel::SharedStorage));
+    size_t smem_size = sizeof(typename Gemm::GemmKernel::SharedStorage);
 
     cudaDeviceProp properties;
     int device_idx;
